WIP test shared buffers resizing and checkpoint
A new test triggers an injection point in the BufferSync() after it has
collected buffers to flushed. Simultaneously it starts buffer shrinking. The
expectation is that the checkpointer would crash accessing a buffer (descriptor)
outside the new range of shared buffers. But that does not happen because of a
bug in synchronization. The checkpointer does not reload configuration when
checkpoint is going on. It does not load the new value of the configuration.
When the resizing is triggered by the PM, checkpointer receives the proc signal
barrier but it does not start it doesn't enter the barrier mechanism and doesn't
alter its address maps or memory sizes. Hence the test does not crash. But of
course it means that it won't consider the correct size of buffers next time it
performs a checkpoint.

The test was at least useful to detect this anomaly. Once we fix the
synchronization issue we should see the crash and then fix the crash.

Author: Ashutosh Bapat

Notes to reviewers
------------------

1. pg_buffercache used a query on pg_settings to fetch the value of the
   number of buffers. That doesn't work anymore because of change in the
   SHOW shared_buffers. Modified the test to convert the setting value
   to the number of shared buffers, save it in a variable and use the
   variable in queries which need the number of shared buffers. We could
   instead fix ShowGUCOption() to pass use_units flag to show_hook and
   let it output the number of shared buffers instead.  But that seems a
   larger change. There aren't other GUCs whose show_hook outputs their
   values with units. So this local fix might be better.